requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-allauth==0.51.0
dj-rest-auth==2.2.5
flake8==5.0.4
isort==5.10.1
pillow==9.3.0
yapf==0.32.0
